requirements.txt
mysql-connector==2.2.9
mysql-connector-python==8.0.19
pandas==1.0.2
pip==20.0.2
wheel==0.34.2
python-dotenv
